exec/backend/requirements.txt
APScheduler==3.7.0
asgiref==3.2.10
branca==0.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.7.1
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
gTTS==2.2.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.17.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pandas==0.25.3
parse==1.19.0
playsound==1.2.2
protobuf==3.15.6
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.3.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
